requirements.txt
python-dotenv
boto3
adafruit-circuitpython-ads1x15
mysql-connector
picamera